requirements.txt

apache-airflow-providers-google

pandas
numpy
pyarrow
pytrends

urllib3==1.26.18

google-cloud-storage
google-cloud-bigquery
gcsfs
fsspec

python-dateutil

pytest
pytest-cov

